requirements.txt
fastapi
uvicorn[standard]
pydantic
prometheus-client
torch
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
pynvml==11.5.0
sse-starlette
redis>=4.2.0
protobuf==6.31.1
grpcio==1.76.0
grpcio-tools==1.76.0

